requirements.txt
# requirements.txt

pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.3.0
flake8>=4.0.1
mypy>=0.950
pytest-cov>=3.0.0